Fix horizontal occsional alingment issues in OpenGL cached glyphs
The left edge of the src location rectangle for the first glyph in a
cache chunk would set to extraspace and not zero, but all other
rectangles in the chunk would be set from the aEnds array. This produced
a bug where only certain letters would be mispositioned, proportional to
the fonts point size.

Change-Id: Id53b4a13d6d1534fc189fcd05ade77f1b7f90578