Implement processing of (all) current LO tasks
Processing all current events is currently just implemented for
LO user events and system events, which is inconsistent. This
implements the same functionality in the scheduler to process all
current LO tasks.

Currently used by DoYield in the Mac and Windows backends.

This also fixes the testTriggerIdleFromIdle test, as this now
correctly wait-yields until the watchdog kicks in.

Change-Id: Icab8671ceae724a96959fa092ad00bff20c919e1